+++ contracts/mocks/account/AccountERC7702Mock.sol
@@ -3,6 +3,6 @@
 pragma solidity ^0.8.20;
 
 import {Account} from "../../account/Account.sol";
-import {AccountSignerERC7702} from "../../account/extensions/AccountSignerERC7702.sol";
+import {SignerERC7702} from "../../utils/cryptography/SignerERC7702.sol";
 
-abstract contract AccountERC7702Mock is Account, AccountSignerERC7702 {}
+abstract contract AccountERC7702Mock is Account, SignerERC7702 {}