+++ test/account/AccountERC7702.test.js
@@ -3,12 +3,9 @@ const { loadFixture } = require('@nomicfoundation/hardhat-network-helpers');
 const { ERC4337Helper } = require('../helpers/erc4337');
 const { PackedUserOperation } = require('../helpers/eip712-types');
 
-const {
-  shouldBehaveLikeAccountCore,
-  shouldBehaveLikeAccountERC7821,
-  shouldBehaveLikeAccountHolder,
-} = require('./Account.behavior');
+const { shouldBehaveLikeAccountCore, shouldBehaveLikeAccountHolder } = require('./Account.behavior');
 const { shouldBehaveLikeERC7739 } = require('../utils/cryptography/ERC7739.behavior');
+const { shouldBehaveLikeERC7821 } = require('./extensions/ERC7821.behavior');
 
 async function fixture() {
   // EOAs and environment
@@ -45,8 +42,8 @@ describe('AccountERC7702', function () {
   });
 
   shouldBehaveLikeAccountCore();
-  shouldBehaveLikeAccountERC7821({ deployable: false });
   shouldBehaveLikeAccountHolder();
+  shouldBehaveLikeERC7821({ deployable: false });
 
   describe('ERC7739', function () {
     beforeEach(async function () {