+++ CHANGELOG.md
@@ -12,7 +12,7 @@
 - `AccountCore`: Added a simple ERC-4337 account implementation with minimal logic to process user operations.
 - `Account`: Extensions of AccountCore with recommended features that most accounts should have.
 - `AbstractSigner`, `SignerECDSA`, `SignerP256`, and `SignerRSA`: Add an abstract contract, and various implementations, for contracts that deal with signature verification. Used by AccountCore and `ERC7739Utils.
-- `AccountSignerERC7702`: Implementation of `AbstractSigner` for ERC-7702 compatible accounts.
+- `SignerERC7702`: Implementation of `AbstractSigner` for Externally Owned Accounts (EOAs). Useful with ERC-7702.
 
 ## 13-12-2024